Renamed DetailsDataToolbar as DetailsToolbar
